refactor: move src/refinements to src/schemas/refinements
