+++ src/schemas/sdo/group.schema.ts
@@ -1,11 +1,11 @@
-import { z } from 'zod/v4';
 import { attackBaseDomainObjectSchema } from '@/schemas/common/attack-base-object.js';
 import { createStixTypeValidator } from '@/schemas/common/stix-type.js';
-import { createFirstAliasRefinement } from '@/refinements/index.js';
+import { createFirstAliasRefinement } from '@/schemas/refinements/index.js';
+import { z } from 'zod/v4';
 import {
   aliasesSchema,
-  createStixIdValidator,
   createAttackExternalReferencesSchema,
+  createStixIdValidator,
   stixTimestampSchema,
   xMitreDomainsSchema,
   xMitreModifiedByRefSchema,
