+++ src/schemas/sdo/malware.schema.ts
@@ -1,6 +1,8 @@
+import {
+  createFirstAliasRefinement,
+  createFirstXMitreAliasRefinement,
+} from '@/schemas/refinements/index.js';
 import { z } from 'zod/v4';
-import { createStixTypeValidator } from '../common/stix-type.js';
-import { softwareSchema } from './software.schema.js';
 import {
   createAttackExternalReferencesSchema,
   createOldMitreAttackIdSchema,
@@ -9,15 +11,13 @@ import {
   stixTimestampSchema,
 } from '../common/index.js';
 import {
-  MalwareCapabilityOV,
-  ProcessorArchitectureOV,
   ImplementationLanguageOV,
+  MalwareCapabilityOV,
   MalwareTypeOV,
+  ProcessorArchitectureOV,
 } from '../common/open-vocabulary.js';
-import {
-  createFirstAliasRefinement,
-  createFirstXMitreAliasRefinement,
-} from '@/refinements/index.js';
+import { createStixTypeValidator } from '../common/stix-type.js';
+import { softwareSchema } from './software.schema.js';
 
 /////////////////////////////////////
 //
