+++ src/schemas/sdo/stix-bundle.schema.ts
@@ -1,29 +1,29 @@
+import { createFirstBundleObjectRefinement } from '@/schemas/refinements/index.js';
 import { z } from 'zod/v4';
-import { createStixTypeValidator } from '../common/stix-type.js';
 import { createStixIdValidator } from '../common/stix-identifier.js';
-import { type Malware, malwareSchema } from './malware.schema.js';
-import { type Asset, assetSchema } from './asset.schema.js';
-import { type Campaign, campaignSchema } from './campaign.schema.js';
-import { type DataComponent, dataComponentSchema } from './data-component.schema.js';
-import { type LogSource, logSourceSchema } from './log-source.schema.js';
-import { type DataSource, dataSourceSchema } from './data-source.schema.js';
-import { type Identity, identitySchema } from './identity.schema.js';
-import { type Matrix, matrixSchema } from './matrix.schema.js';
-import { type Tool, toolSchema } from './tool.schema.js';
-import { type Tactic, tacticSchema } from './tactic.schema.js';
-import { type Technique, techniqueSchema } from './technique.schema.js';
-import { type Group, groupSchema } from './group.schema.js';
-import { type Mitigation, mitigationSchema } from './mitigation.schema.js';
-import { type Collection, collectionSchema } from './collection.schema.js';
-import { type DetectionStrategy, detectionStrategySchema } from './detection-strategy.schema.js';
-import { type Analytic, analyticSchema } from './analytic.schema.js';
-import { type Relationship, relationshipSchema } from '../sro/relationship.schema.js';
+import { type StixSpecVersion, stixSpecVersionSchema } from '../common/stix-spec-version.js';
+import { createStixTypeValidator } from '../common/stix-type.js';
 import {
   type MarkingDefinition,
   markingDefinitionSchema,
 } from '../smo/marking-definition.schema.js';
-import { createFirstBundleObjectRefinement } from '@/refinements/index.js';
-import { stixSpecVersionSchema, type StixSpecVersion } from '../common/stix-spec-version.js';
+import { type Relationship, relationshipSchema } from '../sro/relationship.schema.js';
+import { type Analytic, analyticSchema } from './analytic.schema.js';
+import { type Asset, assetSchema } from './asset.schema.js';
+import { type Campaign, campaignSchema } from './campaign.schema.js';
+import { type Collection, collectionSchema } from './collection.schema.js';
+import { type DataComponent, dataComponentSchema } from './data-component.schema.js';
+import { type DataSource, dataSourceSchema } from './data-source.schema.js';
+import { type DetectionStrategy, detectionStrategySchema } from './detection-strategy.schema.js';
+import { type Group, groupSchema } from './group.schema.js';
+import { type Identity, identitySchema } from './identity.schema.js';
+import { type LogSource, logSourceSchema } from './log-source.schema.js';
+import { type Malware, malwareSchema } from './malware.schema.js';
+import { type Matrix, matrixSchema } from './matrix.schema.js';
+import { type Mitigation, mitigationSchema } from './mitigation.schema.js';
+import { type Tactic, tacticSchema } from './tactic.schema.js';
+import { type Technique, techniqueSchema } from './technique.schema.js';
+import { type Tool, toolSchema } from './tool.schema.js';
 
 export type AttackObject =
   | Malware
