+++ test/objects/data-source.test.ts
@@ -1,9 +1,9 @@
 import { describe, expect, it } from 'vitest';
-import { createSyntheticStixObject } from '../../src/generator';
 import {
-    type ExternalReferences
+  type ExternalReferences
 } from '../../src/schemas/common/index';
 import { type DataSource, dataSourceSchema } from '../../src/schemas/sdo/data-source.schema';
+import { createSyntheticStixObject } from '../../src/utils/index';
 
 describe('dataSourceSchema', () => {
   const minimalDataSource = createSyntheticStixObject('x-mitre-data-source');
