+++ test/objects/detection-strategy.test.ts
@@ -1,10 +1,10 @@
 import { v4 as uuidv4 } from 'uuid';
 import { describe, expect, it } from 'vitest';
-import { createSyntheticStixObject } from '../../src/generator';
 import {
-    type ExternalReferences
+  type ExternalReferences
 } from '../../src/schemas/common/index';
 import { type DetectionStrategy, detectionStrategySchema } from '../../src/schemas/sdo/detection-strategy.schema';
+import { createSyntheticStixObject } from '../../src/utils/index';
 
 describe('detectionStrategySchema', () => {
   const minimalDetectionStrategy = createSyntheticStixObject('x-mitre-detection-strategy');
