+++ test/objects/log-source.test.ts
@@ -1,10 +1,10 @@
 import { v4 as uuidv4 } from 'uuid';
 import { describe, expect, it } from 'vitest';
-import { createSyntheticStixObject } from '../../src/generator';
 import {
-    type ExternalReferences
+  type ExternalReferences
 } from '../../src/schemas/common/index';
 import { type LogSource, logSourceSchema } from '../../src/schemas/sdo/log-source.schema';
+import { createSyntheticStixObject } from '../../src/utils/index';
 
 describe('logSourceSchema', () => {
   const minimalLogSource = createSyntheticStixObject('x-mitre-log-source');
