+++ test/objects/malware.test.ts
@@ -1,9 +1,9 @@
 import { describe, expect, it } from 'vitest';
-import { createSyntheticStixObject } from '../../src/generator';
 import {
-    type StixTimestamp
+  type StixTimestamp
 } from '../../src/schemas/common/index';
 import { type Malware, malwareSchema } from '../../src/schemas/sdo/malware.schema';
+import { createSyntheticStixObject } from '../../src/utils/index';
 
 describe('MalwareSchema', () => {
   const minimalMalware = createSyntheticStixObject('malware');
